requirements.txt
win11toast==0.35
flask==3.0.3
pywebview==5.1
google-api-python-client
google-auth-httplib2
google-auth-oauthlib